% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators.R
\name{attack_eff}
\alias{attack_eff}
\alias{serve_eff}
\alias{reception_eff}
\title{Various skill performance indicators}
\usage{
attack_eff(evaluation, skill)

serve_eff(evaluation, skill)

reception_eff(evaluation, skill)
}
\arguments{
\item{evaluation}{character: vector of skill evaluations ("Winning attack", "Error", etc)}

\item{skill}{character: (optional) vector of skill values ("Attack", "Block", etc). If provided, it will be used to filter the \code{evaluation} vector to the elements corresponding to the correct skill. If not provided, all elements of \code{evaluation} will be used}
}
\value{
A numeric scalar
}
\description{
\itemize{
\item attack_eff: (number of kills - number of errors and blocked attacks) / (number of attacks)
\item serve_eff: (number of aces and positive serves - number of errors and poor serves) / (number of serves)
\item reception_eff: (number of perfect and positive passes - number of errors and overpasses) / (number of passes)
}
}
\examples{
\dontrun{
 library(dplyr)
 x <- ovdata_example("mlafin_braslovce_nkbm", as = "parsed")
 plays(x) \%>\% dplyr::filter(skill == "Attack") \%>\% group_by(player_name) \%>\%
   dplyr::summarize(N_attacks = n(), att_eff = attack_eff(evaluation))
}

}
