% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attack_evenness.R
\name{ov_aev_reference_props}
\alias{ov_aev_reference_props}
\title{Reference proportions used in attack evenness calculations}
\usage{
ov_aev_reference_props()
}
\value{
A tibble with columns \code{player_role} ("middle", "outside", "opposite", "setter", "libero"), \code{player_front_back} (whether that player is front or back court), and \code{p_expected} (the proportion of attacks that the player is expected to make under a perfectly even attack scenario)
}
\description{
Reference proportions used in attack evenness calculations
}
\seealso{
\code{\link[=ov_aev]{ov_aev()}}
}
