% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liberos.R
\name{ov_augment_liberos}
\alias{ov_augment_liberos}
\title{Add information about liberos to a datavolley plays dataframe}
\usage{
ov_augment_liberos(x, liberos, middles)
}
\arguments{
\item{x}{data.frame: the \code{plays} data.frame as returned by \code{\link[datavolley:dv_read]{datavolley::read_dv()}} or \code{\link[peranavolley:pv_read]{peranavolley::pv_read()}}. Can be \code{plays} from multiple matches combined using \code{rbind} or \code{bind_rows}}

\item{liberos}{character: vector of \code{player_id} values of liberos. Or a data.frame with \code{match_id}, \code{player_id}, \code{team}, \code{role} columns (only rows with \code{role} == "libero" will be used as liberos). The latter form will be necessary if \code{x} contains data from multiple matches and a player plays as a libero in some matches but other positions in other matches}

\item{middles}{character: vector of \code{player_id} values of middle blockers. Or a data.frame with \code{match_id}, \code{player_id}, \code{team}, \code{role} columns (only rows with \code{role} == "middle" will be used as middles)}
}
\value{
A data.frame of x augmented with columns \code{libero_id} and \code{libero_on_court}. See Details for the contents of these columns
}
\description{
Add information about liberos to a datavolley plays dataframe
}
\details{
DataVolley (\code{.dvw}) files do not include information about the libero on court. This function adds that information where it can. It assumes that the libero replaces the back-court middle blocker, except when a middle blocker is serving.

The procedure is broadly:
\itemize{
\item all recorded ball touches are checked on a per-set basis for each team: if there were no ball touches by any libero (for a given team) during a set, it is assuumed that no libero was used in that set by that team. If ball touches were recorded only for one libero, it is assumed that only a single libero was used by that team in that set, even if that team has multiple liberos available
\item this information is then refined on a rally-by-rally basis: in a rally where the libero made a ball touch we can be sure that this libero was on court
}

The function returns a copy of \code{x} with two columns added:
\itemize{
\item the \code{libero_on_court} column will contain TRUE if a libero was on court for that data row, or FALSE if not on court, or NA if unknown
\item the \code{libero_id} column will contain the \code{player_id} of the libero that was on court for that data row. If no libero was on court, the value "\emph{NO_LIBERO}" will be used. If a libero was known to be on court but we could not identify who it was, the value "\emph{MULTIPLE_LIBEROS}" will be used
}
}
\examples{
x <- ovdata_example("190301_kats_beds", as = "parsed")

## identify our liberos and middles
liberos <- c(x$meta$players_h$player_id[which(x$meta$players_h$role == "libero")],
             x$meta$players_v$player_id[which(x$meta$players_v$role == "libero")])
middles <- c(x$meta$players_h$player_id[which(x$meta$players_h$role == "middle")],
             x$meta$players_v$player_id[which(x$meta$players_v$role == "middle")])

## infer libero info
px <- ov_augment_liberos(plays(x), liberos = liberos, middles = middles)

}
