% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment.R
\name{ov_augment_plays}
\alias{ov_augment_plays}
\title{Add some extra columns to a plays object}
\usage{
ov_augment_plays(
  x,
  to_add = c("receiving_team", "touch_summaries", "setters"),
  use_existing = TRUE
)
}
\arguments{
\item{x}{data.frame: the \code{plays} data.frame as returned by \code{\link[datavolley:read_dv]{datavolley::read_dv()}} or \code{\link[peranavolley:pv_read]{peranavolley::pv_read()}}}

\item{to_add}{character: columns to add
\itemize{
\item "receiving_team" adds the receiving team name
\item "touch_summaries" adds a number of columns named "ts_*" that summarize a team touch (e.g. columns "ts_pass_quality", "ts_pass_evaluation_code" give the pass quality and pass evaluation code of the reception or dig associated with a given team touch)
\item "followed" adds the columns "followed_timeout", "followed_technical_timeout", and "followed_sub"
\item "setters" adds the columns "home_setter_id", "visiting_setter_id" (the player IDs of the home and visiting setter on court), and "setter_id" and "setter_position" (the player ID and position of the setter from the team performing the current action)
}}

\item{use_existing}{logical: if \code{TRUE} and all of the columns associated with a given \code{to_add} choice are already present in \code{x}, they won't be re-generated}
}
\value{
\code{x} with the extra columns added
}
\description{
Add some extra columns to a plays object
}
