% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setter_choice_functions.R
\name{ov_create_history_table}
\alias{ov_create_history_table}
\title{Create a prior table from a dvw or a directory of dvw files}
\usage{
ov_create_history_table(
  dvw,
  play_phase = c("Reception", "Transition"),
  attack_by = "code",
  setter_position_by = "rotation",
  exclude_attacks = c("PR"),
  normalize_parameters = TRUE
)
}
\arguments{
\item{dvw}{string: path to one or more datavolley files, a list of one or more datavolley objects, or a directory containing datavolley files}

\item{play_phase}{character: one or both of "Reception", "Transition"}

\item{attack_by}{string: either "code", "zone", "tempo" or "setter call"}

\item{setter_position_by}{string: either "rotation", or "front_back"}

\item{exclude_attacks}{character: vector of attack codes to exclude}

\item{normalize_parameters}{logical: reduce the prior parameter values}
}
\value{
A list, currently with one component named "prior_table"
}
\description{
Create a prior table from a dvw or a directory of dvw files
}
\examples{
## use this file to create the priors
hist_dvw <- ovdata_example("NCA-CUB")
history_table <- ov_create_history_table(dvw = hist_dvw, attack_by = "setter call",
                                   setter_position_by = "front_back")

## use it on another file (here, the same file for demo purposes)
## usually the history would be from a reference set of previous matches

dvw <- ovdata_example("NCA-CUB")
setter <- ov_simulate_setter_distribution(dvw = dvw, play_phase = "Reception", n_sim = 100,
                                  attack_by = "setter call", attack_options = "use_history",
                                  setter_position_by = "front_back",
                                  history_table = history_table, filter_sim = TRUE)

## plot the results
ov_plot_ssd(setter, overlay_set_number = TRUE)
ov_plot_distribution(setter)

}
