% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_files.R
\name{ov_example_file}
\alias{ov_example_file}
\title{Example DataVolley files provided as part of the ovlytics package}
\usage{
ov_example_file(choice = "190301_kats_beds")
}
\arguments{
\item{choice}{string: which data file to return?
\itemize{
\item "190301_kats_beds" - a match between GKS Katowice and MKS Bedzin during the 2018/19 Polish Plus Liga
}}
}
\value{
path to the file
}
\description{
Example DataVolley files provided as part of the ovlytics package
}
\examples{
myfile <- ov_example_file()
x <- dv_read(myfile)
summary(x)

}
