% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setter_choice_functions.R
\name{ov_plot_history_table}
\alias{ov_plot_history_table}
\title{Plot the prior table}
\usage{
ov_plot_history_table(history_table, team, setter_id)
}
\arguments{
\item{history_table}{data.frame: the \code{prior_table} component of the object returned by \code{\link[=ov_create_history_table]{ov_create_history_table()}}}

\item{team}{string: team name}

\item{setter_id}{string: setter_id}
}
\description{
Plot the prior table
}
\examples{
hist_dvw <- ovdata_example("NCA-CUB")
history_table <- ov_create_history_table(dvw = hist_dvw, setter_position_by = "front_back",
                                                    normalize_parameters = FALSE)
team = unique(history_table$prior_table$team)[1]
setter_id = unique(history_table$prior_table$setter_id)[1]
ov_plot_history_table(history_table, team, setter_id)
}
