% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setter_choice_functions.R
\name{ov_plot_rate}
\alias{ov_plot_rate}
\title{Plot the rates}
\usage{
ov_plot_rate(ssd, team, setter_id, range = c(0.05, 0.95))
}
\arguments{
\item{ssd}{simulated setter distribution output as returned by \code{\link[=ov_simulate_setter_distribution]{ov_simulate_setter_distribution()}}}

\item{team}{string: team name}

\item{setter_id}{string: setter_id}

\item{range}{vector of maximum and minimum quantile value description}
}
\description{
Plot the rates
}
\examples{
dvw <- ovdata_example("NCA-CUB")
system.time({
  ssd <- ov_simulate_setter_distribution(dvw = dvw, play_phase = c("Reception", "Transition"),
                                         n_sim = 150, setter_position_by = "front_back", 
                                          attack_by = "zone")
  team <- ssd$raw_data$meta$teams$team[1]
  setter_id <- ssd$raw_data$meta$players_h$player_id[which(ssd$raw_data$meta$players_h$role == "setter")][2]
  ov_plot_rate(ssd, team, setter_id)
})
}
