% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setter_choice_functions.R
\name{ov_plot_sequence_distribution}
\alias{ov_plot_sequence_distribution}
\title{Plot a simulated setter distribution sequence}
\usage{
ov_plot_sequence_distribution(
  ssd,
  label_setters_by = "id",
  font_size = 11,
  title_wrap = NA,
  split_set = FALSE,
  output = "plot"
)
}
\arguments{
\item{ssd}{simulated setter distribution output as returned by \code{\link[=ov_simulate_setter_distribution]{ov_simulate_setter_distribution()}}}

\item{label_setters_by}{string: either "id" or "name"}

\item{font_size}{numeric: font size}

\item{title_wrap}{numeric: if non-\code{NA}, use \code{\link[=strwrap]{strwrap()}} to break the title into lines of this width}

\item{split_set}{boolean: if \code{TRUE}, separate the distribution sequence by set}

\item{output}{string: either "plot" or "list"}
}
\description{
Plot a simulated setter distribution sequence
}
\examples{
dvw <- ovdata_example("190301_kats_beds")
ssd <- ov_simulate_setter_distribution(dvw = dvw, play_phase = c("Reception"),
                                       n_sim = 100, attack_by = "zone")
ov_plot_sequence_distribution(ssd)

}
