% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setter_choice_functions.R
\name{ov_plot_ssd}
\alias{ov_plot_ssd}
\title{Plot a simulated setter distribution}
\usage{
ov_plot_ssd(
  ssd,
  overlay_set_number = FALSE,
  label_setters_by = "name",
  font_size = 11
)
}
\arguments{
\item{ssd}{simulated setter distribution output as returned by \code{\link[=ov_simulate_setter_distribution]{ov_simulate_setter_distribution()}}}

\item{overlay_set_number}{boolean: if \code{TRUE}, overlay set number and score in the plot}

\item{label_setters_by}{string: either "id" or "name"}

\item{font_size}{numeric: font size}
}
\description{
Plot a simulated setter distribution
}
\examples{
dvw <- ovdata_example("190301_kats_beds")
setter <- ov_simulate_setter_distribution(dvw = dvw, 
                                          n_sim = 150, attack_by = "zone")
ov_plot_ssd(setter, overlay_set_number = TRUE)
}
