% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setter_choice_functions.R
\name{ov_print_rate_table}
\alias{ov_print_rate_table}
\title{Print the rate table}
\usage{
ov_print_rate_table(ssd, team, setter_id)
}
\arguments{
\item{ssd}{simulated setter distribution output as returned by \code{\link[=ov_simulate_setter_distribution]{ov_simulate_setter_distribution()}}}

\item{team}{string: team name}

\item{setter_id}{string: setter_id}
}
\description{
Print the rate table
}
\examples{
dvw <- ovdata_example("NCA-CUB")
system.time({
  ssd <- ov_simulate_setter_distribution(dvw = dvw, play_phase = "Reception",
                                         n_sim = 500, setter_position_by = "front_back")
  team <- ssd$raw_data$meta$teams$team[1]
  setter_id <- ssd$raw_data$meta$players_h$player_id[which(ssd$raw_data$meta$players_h$role == "setter")][1]
  ov_print_rate_table(ssd, team, setter_id)
})
}
