% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/season_table.R
\name{ov_season_table}
\alias{ov_season_table}
\title{Create a summary table of a team's matches in a season}
\usage{
ov_season_table(xl, target_team, target_team_id, show_by = "match date")
}
\arguments{
\item{xl}{list: list of datavolley objects (each as returned by \code{\link[datavolley:read_dv]{datavolley::dv_read()}}}

\item{target_team}{string: the name of the target team. Only one of \code{target_team} or \code{target_team_id} is required}

\item{target_team_id}{string: the team ID of the target team. Ignored if \code{target_team} has been provided}

\item{show_by}{string: either "match date" (show each match according to its date) or "filename" (show each match according to its filename. This might be useful if the match dates are being parsed incorrectly by \code{\link[datavolley:read_dv]{datavolley::dv_read()}})}
}
\value{
A tibble with columns "Opponent", "Date" (or "File"), "Result", "Set scores", and one column for sets 1 to 5
}
\description{
Create a summary table of a team's matches in a season
}
\examples{
## trivial example of a single-match "season"
library(datavolley)
x <- dv_read(dv_example_file())
ov_season_table(list(x), target_team = home_team(x))

}
