% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setter_choice_functions.R
\name{ov_simulate_multiple_setter_distribution}
\alias{ov_simulate_multiple_setter_distribution}
\title{Simulate a Bayesian Bandit choice for a given set of probabilities and a number of points for multiple games}
\usage{
ov_simulate_multiple_setter_distribution(
  list_dv,
  play_phase = c("Reception", "Transition"),
  n_sim = 500,
  priors = list(name = "beta", par1 = 1, par2 = 1),
  epsilon = 1,
  filter_sim = FALSE,
  attack_options = "use_data",
  setter_position_by = "rotation",
  history_table = NULL,
  attack_by = "code",
  exclude_attacks = c("PR"),
  rotation = "SHM",
  shiny_progress = NULL
)
}
\arguments{
\item{list_dv}{list: list of datavolley object as returned by \code{\link[datavolley:dv_read]{datavolley::dv_read()}}}

\item{play_phase}{character: one or both of "Reception", "Transition"}

\item{n_sim}{integer: number of simulations}

\item{priors}{numeric: prior distribution of the kill rate for the different attacking options}

\item{epsilon}{numeric: reward size}

\item{filter_sim}{logical:}

\item{attack_options}{string: either "use_data" or "use_history"}

\item{setter_position_by}{string: either "rotation" or "front_back"}

\item{history_table}{list: (only if \code{attack_options} is "use_history") the object returned by \code{\link[=ov_create_history_table]{ov_create_history_table()}}}

\item{attack_by}{string: either "code", "zone", "tempo", "setter call", "attacker_name", "player_role"}

\item{exclude_attacks}{character: vector of attack codes to exclude}

\item{rotation}{string: (only relevant when \code{attack_by} is "player_role") either "SHM" (assume a setter-hitter-middle rotation order), or "SMH" (setter-middle-hitter)}

\item{shiny_progress}{numeric: an optional two-element vector. If not \code{NULL} or \code{NA}, \code{\link[shiny:withProgress]{shiny::setProgress()}} calls will be made during simulation with \code{value}s in this range}
}
\description{
Simulate a Bayesian Bandit choice for a given set of probabilities and a number of points for multiple games
}
\examples{
list_dv <- list(dv_read(ovdata_example("NCA-CUB")))
system.time({
  mssd <- ov_simulate_multiple_setter_distribution(list_dv = list_dv, play_phase = "Reception",
               n_sim = 100, setter_position_by = "front_back")
})
}
\seealso{
\code{\link[=ov_simulate_setter_distribution]{ov_simulate_setter_distribution()}}
}
