% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setter_choice_functions.R
\name{ov_simulate_setter_distribution}
\alias{ov_simulate_setter_distribution}
\title{Simulate a Bayesian Bandit choice for a given set of probabilities and a number of points}
\usage{
ov_simulate_setter_distribution(
  dvw,
  play_phase = c("Reception", "Transition"),
  n_sim = 500,
  priors = list(name = "beta", par1 = 1, par2 = 1),
  epsilon = 1,
  filter_sim = FALSE,
  attack_options = "use_data",
  setter_position_by = "rotation",
  history_table = NULL,
  attack_by = "code",
  exclude_attacks = c("PR"),
  shiny_progress = NULL
)
}
\arguments{
\item{dvw}{string or datavolley: a datavolley object as returned by \code{\link[datavolley:read_dv]{datavolley::dv_read()}} or a path to datavolley file}

\item{play_phase}{character: one or both of "Reception", "Transition"}

\item{n_sim}{integer: number of simulations}

\item{priors}{numeric: prior distribution of the kill rate for the different attacking options}

\item{epsilon}{numeric: reward size}

\item{filter_sim}{logical:}

\item{attack_options}{string: either "use_data" or "use_history"}

\item{setter_position_by}{string: either "rotation" or "front_back"}

\item{history_table}{list: (only if \code{attack_options} is "use_history") the object returned by \code{\link[=ov_create_history_table]{ov_create_history_table()}}}

\item{attack_by}{string: either "code", "zone", "tempo", "setter call"}

\item{exclude_attacks}{character: vector of attack codes to exclude}

\item{shiny_progress}{numeric: an optional two-element vector. If not \code{NULL} or \code{NA}, \code{\link[shiny:withProgress]{shiny::setProgress()}} calls will be made during simulation with \code{value}s in this range}
}
\description{
Simulate a Bayesian Bandit choice for a given set of probabilities and a number of points
}
\examples{
dvw <- ovdata_example("190301_kats_beds")
system.time({
  ssd <- ov_simulate_setter_distribution(dvw = dvw, play_phase = "Reception",
                                         n_sim = 100, attack_by = "setter call", 
                                         setter_position_by = "rotation")
})
}
\seealso{
\code{\link[=ov_create_history_table]{ov_create_history_table()}}
}
