% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{ov_sort_attack_codes}
\alias{ov_sort_attack_codes}
\title{Sort DataVolley attack codes}
\usage{
ov_sort_attack_codes(ac, by = "XV", na.last = NA)
}
\arguments{
\item{ac}{character: character vector of attack codes to sort}

\item{by}{string: method to use, currently only "XV" (any other value will default back to using \code{\link{sort}} without modification). "XV" will place X and V codes first (in numerical order, with each X preceding its matching V) then everything else in alphabetical order after that}

\item{na.last}{logical: passed to \code{sort}}
}
\value{
Sorted character vector
}
\description{
Sort DataVolley attack codes
}
\examples{
ov_sort_attack_codes(c("V5", "V1", "X6", "CF", "X5"))

\dontrun{
  ## sorting might be useful for controlling the plot order when facetting
  ##  a `ggplot` by attack code
  mydata$attack_code <- factor(mydata$attack_code,
            levels = ov_sort_attack_codes(unique(na.omit(mydata$attack_code))))
  ggplot(mydata, ...) + facet_wrap(~attack_code)
}

}
