% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re_exports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{dv_read}
\alias{ovdata_example}
\alias{plays}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{datavolley}{\code{\link[datavolley:read_dv]{dv_read}}, \code{\link[datavolley]{plays}}}

  \item{dplyr}{\code{\link[dplyr:reexports]{\%>\%}}}

  \item{ovdata}{\code{\link[ovdata]{ovdata_example}}}
}}

