% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setter_choice_functions.R
\name{ov_create_history_table}
\alias{ov_create_history_table}
\title{Create a prior table from a dvw or a directory of dvw files}
\usage{
ov_create_history_table(
  dvw,
  play_phase = c("Reception", "Transition"),
  attack_by = "code",
  exclude_attacks = c("PR")
)
}
\arguments{
\item{dvw}{string: path to one or more datavolley files, a list of one or more datavolley objects, or a directory containing datavolley files}

\item{play_phase}{character: one or both of "Reception", "Transition"}

\item{attack_by}{string: either "code" or "zone"}

\item{exclude_attacks}{character: vector of attack codes to exclude}
}
\value{
A list, currently with one component named "prior_table"
}
\description{
Create a prior table from a dvw or a directory of dvw files
}
\examples{
## use this file to create the priors
hist_dvw <- ovdata_example("PM06")
history_t <- ov_create_history_table(dvw = hist_dvw, attack_by = "zone")

## use it on this file
dvw <- ovdata_example("190301_kats_beds")
setter <- ov_simulate_setter_distribution(dvw = dvw, play_phase = "Reception", n_sim = 500,
                                  attack_by = "zone", attack_options = "use_history",
                                  history_table = history_t$prior_table, filter_sim = TRUE)

## plot the results
ov_plot_ssd(setter, overlay_set_number = TRUE)
ov_plot_distribution(setter)

}
