% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setting.R
\name{ov_setter_repetition}
\alias{ov_setter_repetition}
\title{Tabulate setter repeat patterns}
\usage{
ov_setter_repetition(
  x,
  setter_id,
  setter_name,
  exclude_attacks = c("PP", "PR", "P2")
)
}
\arguments{
\item{x}{data.frame: the \code{plays} data.frame as returned by \code{\link[datavolley:read_dv]{datavolley::read_dv()}} or \code{\link[peranavolley:pv_read]{peranavolley::pv_read()}}}

\item{setter_id}{string: (optional) the player ID of the setter to analyze (or provide \code{setter_name}). If neither \code{setter_id} nor \code{setter_name} are provided, all setters will be analyzed}

\item{setter_name}{string: (optional) the name of the setter to analyze (ignored if \code{setter_id} is provided). If neither \code{setter_id} nor \code{setter_name} are provided, all setters will be analyzed}

\item{exclude_attacks}{character: vector of attack codes to exclude}
}
\value{
A data.frame with columns "team", "setter_name", "setter_id", "player_name", "player_id", "category", "opportunities", "repeats", "repeat\%"
}
\description{
Note: analysis is done on the basis of attack actions, and simply assumes that the setter on court made the set.
}
\examples{

x <- plays(ovdata_example("190301_kats_beds", as = "parsed"))
set_reps <- ov_setter_repetition(x, setter_name = "Lukas Tichacek")

library(ggplot2)
ggplot(set_reps, aes(x = player_name, y = `repeat\%`)) + geom_col() +
    geom_text(aes(x = player_name, label = paste0("N=", opportunities)),
              angle = 90, y = 100, hjust = 1, inherit.aes = FALSE) +
    facet_wrap(~category) +
    theme_bw() +
    theme(axis.text.x = element_text(angle = 60, vjust = 1, hjust = 1)) +
    labs(x = NULL, y = "Repeat percentage")

}
