% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ovlytics-package.R
\docType{package}
\name{ovlytics}
\alias{ovlytics}
\title{\pkg{ovlytics}}
\description{
Analytical functions for volleyball analytics, to be used in conjunction with the datavolley and peranavolley packages.
}
